06.MultiAgent/requirements.txt
fastapi
uvicorn
openai
pydantic
streamlit
requests
python-dotenv
qdrant-client
rank-bm25
